Failing spec for #10.
When
* using a single git source _with no ref specified_ with one cookbook and resolving,
* then adding another cookbook from the same source for the purpose of forcing an update,
* then re-resolving which forces an update

When librarian attempts write and bounce the lockfile, the first version of the lockfile it writes contains a duplicate git source.

The reason is that reloading the git source with no ref specified from the lockfile includes the default ref, which makes the source different from a git source with no ref specified.