Better updating of cached git sources.
Make sure to hard-reset to get rid of any potential problems from a previous run. This should clear the way for using checkout safely.

Do not merge remote branches into local branches. Instead, just fetch all branches and tags from the remote and just keep the branches as mere remote branches.

Checkout the SHA directly. No need to open concerns about checking out branches that could possibly be updated.

In the case of a branch being specified on the source, this requires looking up the SHA for the remote branch, rather than for the local branch, because remote branches are not merged into local branches. In the case of a tag or SHA being specified on the source, such a lookup appears unnecessary.

Fixes #44. Fixes #49. Thanks @TylerRick and @bradmontgomery.

Replaces original fixes for #36, #37, and #39.