Include existing manifests' dependencies in resolution.
Fixes #57.

The issue was that, when updating, we start off with a set of manifests and we know their dependencies. But we were not adding these to the starting dependencies before resolving, and the resolver was not picking them during resolution, because it only does that for manifests that the resolver discovers. So if any manifests from a prior resolution declared dependencies of their own, they would be forgotten when performing an update.