Detect conflicts faster in the resolver.
This adds an initial check in the recursive resolution process that fails the resolve early if there are any dependencies in the queue that conflict with any of the already-resolved manifests.

These same conflicts would have been detected anyway, when actually attempting to resolve these dependencies in the queue, but we are just detecting them earlier.